% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAphiaID.R
\name{getCodeList}
\alias{getCodeList}
\title{Get List of Codes}
\usage{
getCodeList(code_type)
}
\value{
A data frame.
}
\description{
Retrieves all the codes for the given code type.
}
\examples{
# get the WoRMS AphiaIDs for ICES assessment species
getCodeList("SpecWoRMS")

}

