% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codeLinks.R
\name{linkCodes}
\alias{linkCodes}
\title{Get Links Between Two Code Types}
\usage{
linkCodes(code_type, linked_code_type, link_type = c("children", "parent"))
}
\arguments{
\item{code_type}{the species name, either in English (default) or Latin.}

\item{linked_code_type}{whether the species name is in Latin.}

\item{link_type}{whether to match the species name as a regular expression.}
}
\value{
A data frame.
}
\description{
Find parent or child linkages between the codes of two code types.
}
\examples{
\donttest{
linkCodes("Pargroup", "PARAM", "children")

}
}
\seealso{
\code{\link{icesVocab-package}} gives an overview of the package.
}
\author{
Colin Millar.
}
