% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vocab_api.R
\name{vocab_api}
\alias{vocab_api}
\title{Build a SAG web service url}
\usage{
vocab_api(service, ...)
}
\arguments{
\item{service}{the name of the service}

\item{...}{name arguments will be added as queries}
}
\value{
a complete url as a character string
}
\description{
utility to build a url with optional query arguments
}
\examples{

vocab_api("hi", bye = 21)
vocab_api("CodeType")
options("icesVocab.showURL"=TRUE)
vocab_api("CodeType")

}
