% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCodeDetail.R
\name{getCodeDetail}
\alias{getCodeDetail}
\title{Get List of Details}
\usage{
getCodeDetail(code_type, code)
}
\arguments{
\item{code_type}{the code type, e.g. SpecWoRMS}

\item{code}{the code, e.g. 101170}
}
\value{
A data frame.
}
\description{
Retrieves all the details for the given code.
}
\examples{
# get the details for species code 101170
getCodeDetail("SpecWoRMS", "101170")

}

