% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCodeDetail.R
\name{getCodeDetail}
\alias{getCodeDetail}
\title{Get Details}
\usage{
getCodeDetail(code_type, code)
}
\arguments{
\item{code_type}{the code type, e.g. SpecWoRMS.}

\item{code}{the code, e.g. 101170.}
}
\value{
A data frame.
}
\description{
Get details for a given code.
}
\examples{
# Species code 101170
getCodeDetail("SpecWoRMS", 101170)

}
\author{
Colin Millar.
}
\seealso{
\code{\link{getCodeTypeList}} and \code{\link{getCodeList}} get code types
and codes of a given type.

\code{\link{icesVocab-package}} gives an overview of the package.
}

