% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCodeList.R
\name{getCodeList}
\alias{getCodeList}
\title{Get Codes}
\usage{
getCodeList(code_type, date = NULL)
}
\arguments{
\item{code_type}{the code type, e.g. SpecWoRMS.}

\item{date}{restrict output to codes modified after a given date in
yyyy-mm-dd format, e.g. "2010-12-01".}
}
\value{
A data frame.
}
\description{
Get all codes of a given code type.
}
\examples{
# Species codes
getCodeList("SpecWoRMS")

}

