% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icesVocab-package.R
\docType{package}
\name{icesVocab-package}
\alias{icesVocab-package}
\alias{icesVocab}
\title{ICES Vocabularies Database Web Services}
\description{
R interface to access the RECO POX web services of the ICES (International
Council for the Exploration of the Sea) Vocabularies database.
}
\details{
\emph{Get full tables:}
\tabular{ll}{
  \code{\link{getCodeTypeList}} \tab all code code types\cr
  \code{\link{getCodeList}}     \tab codes of a given type\cr
  \code{\link{getCodeDetail}}   \tab details for a given code
}
\emph{Look up specific codes:}
\tabular{ll}{
  \code{\link{findAphia}} \tab species code
}
}
\references{
ICES Vocabularies database: \url{http://vocab.ices.dk/}.
}
\author{
Colin Millar and Arni Magnusson.
}
