% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkCodes.R
\name{linkCodes}
\alias{linkCodes}
\title{Get Links Between Two Code Types}
\usage{
linkCodes(code_type, linked_code_type, codes = NULL, ...)
}
\arguments{
\item{code_type}{an ICES Vocabulary code type.}

\item{linked_code_type}{another ICES Vocabulary code type.}

\item{codes}{whether to only match a selection of codes within the first code type.}

\item{...}{for back compatibility.}
}
\value{
A data frame.
}
\description{
Find parent or child linkages between the codes of two code types.
}
\examples{
\dontrun{
x <- linkCodes("Pargroup", "PARAM")

# get all stocks in subarea 4a
# note you would need to cross check with icesSD for
# currently active stocks
linkCodes("ICES_Area", "ICES_StockCode", codes = c("27.4.a"))

# get all areas for cod.27.1-2
linkCodes("ICES_StockCode", "ICES_Area", codes = c("cod.27.1-2"))

get areas for cod.27.21, cod.27.24-32 and cod.27.46a7d20
linkCodes("ICES_StockCode", "ICES_Area", codes = c("cod.27.21", "cod.27.24-32", "cod.27.46a7d20"))
}
}
\seealso{
\code{\link{icesVocab-package}} gives an overview of the package.
}
\author{
Colin Millar.
}
