% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkCodes.R
\name{linkCodes}
\alias{linkCodes}
\title{Get Links Between Two Code Types}
\usage{
linkCodes(code_type, linked_code_type, codes = NULL, ...)
}
\arguments{
\item{code_type}{an ICES Vocabulary code type.}

\item{linked_code_type}{another ICES Vocabulary code type.}

\item{codes}{whether to only match a selection of codes within a code type.}

\item{...}{for back compatibility.}
}
\value{
A data frame.
}
\description{
Find parent or child linkages between the codes of two code types.
}
\examples{
\dontrun{
x <- linkCodes("Pargroup", "PARAM")

x <- linkCodes("ICES_Area", "ICES_StockCode", codes = c("27.1", "27.2"))
}
}
\seealso{
\code{\link{icesVocab-package}} gives an overview of the package.
}
\author{
Colin Millar.
}
